vim: Change python3 ver setting, add nogui version
Instead of having a specific python3 configuration option, we now just
check the args.python variable passed. This can be trivially set using
stdenv.lib.overrideDerivation.

The nogui part required a bit more changes, completely removing all of
the package requirements in the configurable.nix package, splitting it
all out to the definition in all-packages.nix.

There is a new variable, args.X11, that specifies the build inputs for
building Vim with support for X. On systems where X may not be
available, using the nogui version will set args.X11 to an emply list,
causing none of the X dependencies to be built for Vim. The gui option
is set to "no" in the nogui option, but can be manually overriden with
the overrideDerivation, though then "X11" will have to also be manually
set.

There are also facilities to specify your own configure flags and inputs
if desired using the extraFlags and extraInputs arrays. This could let
users enable and disable features as they see fit through overrides and
use platform specific configure flags without adding a maintenance
burden on nixpkgs.